# Optimize via direct socket access with loop lock concurrency protection