Alternative (wrt #224) optimization following @navytux idea to access the socket directly.
While #224 monkey patches the write operation to include a synchronizing lock, this PR monkey patches the loop and includes a loop lock which ensures that an application thread cannot execute a task step while the loop processes callbacks.

Initially, I had hoped that this approach would work for SSL connections, too. But the SSL writes include non thread safe loop operations. Thus, like for #224, SSL access cannot be optimized.

Nevertheless, this approach is simpler and therefore potentially more reliable.