@d-maurer 2023-02-02
@navytux The strange test behavior is partly understood: your (relatively new) test `check_race_external_invalidate_vs_disconnect` occasionally leaves threads behind. Some of those threads are not "daemonic", as a consequence the interpreter shutdown is delayed until those threads stop -- what apparently never happens.

Not yet understood: why GHA shows the output only after the test run is cancelled.

@navytux Would you check your test? At least, all used threads should be "daemonic" -- in order not to interfere with the shutdown. In addition, it might be adequate to let the test fail when the threads are not shut down properly. The same problem (use of non daemonic threads) might affect other of your tests as well.

In my local setup, `zope-testrunner -f --test-path ../ZEO/src -s ZEO -a 1000 -vv -t "external_invalidate_vs_disconnect.*Mapping"` often leaves threads behind and then blocks in the shutdown.

I will try to find out why the threads do not reliably terminate in this branch.