@navytux 2023-02-02
Thanks, Dieter. I will try to look into the test in some time. It is again, sadly, a difficult time here and I'm overloaded. Hopefully it might get better in some time. I appologize for the inconvenience.

P.S. One thing I thought out after seeing https://github.com/zopefoundation/ZEO/pull/224#issuecomment-1410125616 is this: given the speedup is significant I've thought about the whole thing as follows: on Linux threads are full processes. It is just that threads inside the same "process" are processes that share address space. This way before #224 what we have for typical `loadBefore` is 4 switches in between processes:

1. Client thread -> IO thread  via `call_soon_threadsafe` which internally sends wakeup via internal socket.
2. IO thread -> ZEO server process (request)
3. ZEO server -> IO thread (reply)
4. IO thread -> client thread (result)

#224 removes 1, but leaves 2,3 and 4. And even if we remove 1 process switch out of 4 the improvements are significant.

This way it makes sense to see if we can remove all but absolutely necessary processes switches. Looking at the problem from this point of view the other unnecessary switch is 3: we could send directly request to ZEO server from Client thread, and we could receive ZEO server reply also directly by that same client thread avoiding intermediate point on the IO thread.

That would work straightforwardly if we have only one client thread, but the reason IO thread exists at all is because we can have several client threads each issuing requests simultaneously.

So what I thought is this: we can organize execution of asyncio loop in our own LoopExecutor. Then when a client wants to issue server request, it creates corresponding coroutine, and then does the following: check if the loop is already running. If not - start running the loop directly by that client thread  until that loading coroutine completes (and gives the result). If loop is already running - schedule coroutine via call_soon_threadsafe and waits the result via ConcurrentFuture. This way at least one client thread would be able to issue server request directly and receive it also directly also serving other client threads along the way the same way as IO thread currently does. And also we will need to keep on running the IO thread, that would periodically run the same procedure to run the loop, so that from-server invalidations are still received even if there is no client activity. All the above logic could be put into LoopExecutor, and then every place that does the request does just create coro, then loop_exe.run_coro(coro).

The loop will be run at arbitrary thread, but only at one thread at a time. I've quickly checked what is the usage of threading.local in asyncio implementation and I did not see any place which might prevent this scheme to work.

I don't have numbers, but I believe reducing the number of process switches from 3 to 2 could bring even more significant latency reduction. That's why I would consider trying this scheme.

Kirill