@d-maurer 2023-02-02
Kirill Smelkov wrote at 2023-2-2 14:10 -0800:
> ...
>P.S. One thing I thought out after seeing https://github.com/zopefoundation/ZEO/pull/224#issuecomment-1410125616 is this: given the speedup is significant I've thought about the whole thing as follows: on Linux threads are full processes. It is just that threads inside the same "process" are processes that share address space. This way before #224 what we have for typical `loadBefore` is 4 switches in between processes:
>
>1. Client thread -> IO thread  via `call_soon_threadsafe` which internally sends wakeup via internal socket.
>2. IO thread -> ZEO server process (request)
>3. ZEO server -> IO thread (reply)
>4. IO thread -> client thread (result)
>
>#224 removes 1, but leaves 2,3 and 4. And even if we remove 1 process switch out of 4 the improvements are significant.
>
>This way it makes sense to see if we can remove all but absolutely necessary processes switches. Looking at the problem from this point of view the other unnecessary switch is 3: we could send directly request to ZEO server from Client thread, and we could receive ZEO server reply also directly by that same client thread avoiding intermediate point on the IO thread.

Possible in principle but not easy to achieve.
You would make the ZEO communication synchronous: the application
thread sends the request and waits for the result.
This would work with RPC calls **BUT** the server can also send
asynchronous calls **AND** you must guarantee that messages
are processed in sent order.

A further difficulty: long running requests (such as pack).
You would not want that during such calls other requests are blocked.
This means that you cannot use simple synchronous communication;
instead the communication channel must be multiplexed between
different application threads and an additional actor processing
asnychonous messages.

> ...
>So what I thought is this: we can organize execution of asyncio loop in our own LoopExecutor. Then when a client wants to issue server request, it creates corresponding coroutine, and then does the following: check if the loop is already running.

The loop typically synchronously waits in a `select/poll`.
When it does that, it must have a thread which does the waiting.
As you have pointed out for my "lock `_run_once`" idea, the
loop is essentially lways running -- the run just consists in
waits to a large part.

Remember also that the loop must always be prepared to
process asynchronus messages -- even when no application
thread is interested in server interaction.