@d-maurer 2023-02-08
Final `zodbshootout` timings (`zodbshootout --pack --fast -l 5 --include-mapping no --threads shared shootout.conf add store update cold prefetch_cold warm`:
```markdown
| Benchmark                                                    | master      | PR                     | cython PR                 |
|--------------------------------------------------------------|:-----------:|:----------------------:|:--------------------------:|
| {c=2 threads, o=1000} cs: add 1000 objects                   | 894 ms      | 700 ms: 1.28x faster   | 435 ms: 2.05x faster       |
| {c=2 threads, o=1000} cs: store 1000 raw pickles             | 439 ms      | 352 ms: 1.25x faster   | 199 ms: 2.21x faster       |
| {c=2 threads, o=1000} cs: update 1000 objects                | 953 ms      | 787 ms: 1.21x faster   | 597 ms: 1.59x faster       |
| {c=2 threads, o=1000} cs: read 1000 cold objects             | 744 ms      | 524 ms: 1.42x faster   | 493 ms: 1.51x faster       |
| {c=2 threads, o=1000} cs: read 1000 cold prefeteched objects | 335 ms      | 344 ms: 1.03x slower   | 320 ms: 1.04x faster       |
| {c=2 threads, o=1000} cs: write/read 1000 objects            | 1.36 sec    | 1.23 sec: 1.10x faster | 1.19 sec: 1.14x faster     |
| Geometric mean                                               | (ref)       | 1.20x faster           | 1.53x faster               |
```
Unlike previous timings, those above use concurrency 2 (--> `c=2`) which is more realistic for high load than the previous concurrency 1. As a consequence, the timings cannot be compared.

Unfortunately, the speedup dropped considerably. Not sure, whether this is due to the increased concurrency or the stricter protection.