@d-maurer 2023-02-09
Corresponding timings with concurrency 1 (`zodbshootout --pack -c 1 --fast -l 5 --include-mapping no --threads shared shootout.conf add store update cold prefetch_cold warm`):
```markdown
| Benchmark                                                    | master  | PR                   | cython PR              |
|--------------------------------------------------------------|:-------:|:--------------------:|:----------------------:|
| {c=1 threads, o=1000} cs: add 1000 objects                   | 461 ms  | 225 ms: 2.05x faster | 227 ms: 2.03x faster   |
| {c=1 threads, o=1000} cs: store 1000 raw pickles             | 431 ms  | 196 ms: 2.19x faster | 195 ms: 2.21x faster   |
| {c=1 threads, o=1000} cs: update 1000 objects                | 480 ms  | 362 ms: 1.33x faster | 240 ms: 2.00x faster   |
| {c=1 threads, o=1000} cs: read 1000 cold objects             | 455 ms  | 339 ms: 1.34x faster | 326 ms: 1.39x faster   |
| {c=1 threads, o=1000} cs: read 1000 cold prefeteched objects | 174 ms  | 175 ms: 1.01x slower | 165 ms: 1.05x faster   |
| {c=1 threads, o=1000} cs: write/read 1000 objects            | 891 ms  | 708 ms: 1.26x faster | 550 ms: 1.62x faster   |
| Geometric mean                                               | (ref)   | 1.47x faster         | 1.66x faster           |
```
Apparently, the efficiency of the optimization drops significantly with an increasing concurrency. The drop is more pronounced without the `cython` optimization.